Split Scenario.hs
